fix: stabilize package CI checks
